PIR_Pipeline/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pir_pipeline"
version = "1.0.0"
authors = [
    { name = "Arch Systems Inc."}
]
description = "Program Information Report ingestion and linking infrastructure."
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: Microsoft :: Windows",
    "PRIVATE :: Do Not Upload"
]
dependencies = [
    "pywin32", "mysql-connector-python"
]

[project.urls]
Homepage = "https://github.com/HHS/ACF-pir-data"

[project.scripts]
pir-ingest = "pir_pipeline:ingestData.main"
pir-link = "pir_pipeline:linkQuestions.main"
pir-r-packages = "pir_pipeline:installPackages.main"
pir-dashboard = "pir_pipeline:startDashboard.main"
pir-status = "pir_pipeline:pirStatus.main"
pir-setup-db = "pir_pipeline:setupDB.main"

[project.gui-scripts]
pir-setup = "pir_pipeline:pirSetup.main"
pir-create-directories = "pir_pipeline:createDirectories.main"
pir-config-db = "pir_pipeline:configureDB.main"

[tools.setuptools]
include-package-data = true